Poison duration increase - Fix a bug where equipped items are counted towards equipped totals

+++ src/main/java/me/neoblade298/neorogue/equipment/consumables/MinorShieldsPotion.java
@@ -16,7 +16,7 @@ import me.neoblade298.neorogue.session.fight.trigger.TriggerResult;
 
 public class MinorShieldsPotion extends Consumable {
 	private static final String ID = "MinorShieldsPotion";
-	private double shields;
+	private int shields;
 	
 	public MinorShieldsPotion(boolean isUpgraded) {
 		super(ID, "Minor Shields Potion", isUpgraded, Rarity.COMMON, EquipmentClass.CLASSLESS);
